+++ example/android/app/src/main/kotlin/com/smartdialog/example/MainActivity.kt
@@ -1,6 +1,0 @@
-package com.smartdialog.example
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}