+++ lib/src/kit/view_utils.dart
@@ -1,5 +1,8 @@
 import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
+
+import '../helper/dialog_proxy.dart';
 
 class ViewUtils {
   static void addSafeUse(VoidCallback callback) {
@@ -9,6 +12,25 @@ class ViewUtils {
     } else {
       callback();
     }
+  }
+
+  static bool isDarkModel() {
+    if (DialogProxy.contextNavigator == null) {
+      return false;
+    }
+
+    var brightness = Theme.of(DialogProxy.contextNavigator!).brightness;
+    return brightness == Brightness.dark;
+  }
+}
+
+class ThemeStyle {
+  static Color get backgroundColor {
+    return ViewUtils.isDarkModel() ? const Color(0xFF606060) : Colors.black;
+  }
+
+  static Color get textColor {
+    return ViewUtils.isDarkModel() ? Colors.white : Colors.white;
   }
 }