+++ lib/src/widget/default/loading_widget.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_smart_dialog/src/kit/view_utils.dart';
 
 class LoadingWidget extends StatelessWidget {
   const LoadingWidget({Key? key, required this.msg}) : super(key: key);
@@ -11,20 +12,20 @@ class LoadingWidget extends StatelessWidget {
     return Container(
       padding: const EdgeInsets.symmetric(horizontal: 30, vertical: 20),
       decoration: BoxDecoration(
-        color: Colors.black,
+        color: ThemeStyle.backgroundColor,
         borderRadius: BorderRadius.circular(15),
       ),
       child: Column(mainAxisSize: MainAxisSize.min, children: [
         //loading animation
-        const CircularProgressIndicator(
+        CircularProgressIndicator(
           strokeWidth: 3,
-          valueColor: AlwaysStoppedAnimation(Colors.white),
+          valueColor: AlwaysStoppedAnimation(ThemeStyle.textColor),
         ),
 
         //msg
         Container(
           margin: const EdgeInsets.only(top: 20),
-          child: Text(msg, style: const TextStyle(color: Colors.white)),
+          child: Text(msg, style: TextStyle(color: ThemeStyle.textColor)),
         ),
       ]),
     );